refactor: atomic 디자인 적용
