+++ src/components/Atoms/Input/input.stories.jsx
@@ -2,7 +2,7 @@ import React from 'react';
 import { BlockInput, UnderlineInput } from '.';
 
 export default {
-  title: 'Payment/Input',
+  title: 'Atoms/Input',
 };
 
 export const BlockedInputExample = args => <BlockInput {...args} />;
