+++ src/components/Molecules/Card/index.jsx
@@ -7,8 +7,8 @@ import {
   CardOwnerName,
   sizeTable,
 } from './style';
-import { DEFAULT_TEXT } from '../../constant';
-import { CARD_NUMBER_MARK, DATE_SEPARATOR } from '../../constant/mark';
+import { DEFAULT_TEXT } from '../../../constant';
+import { CARD_NUMBER_MARK, DATE_SEPARATOR } from '../../../constant/mark';
 import { CARD_SIZE } from 'constant/';
 
 function Card({
