+++ src/components/Molecules/SecurityNumberInput/SecurityNumberInput.stories.jsx
@@ -2,12 +2,12 @@ import React from 'react';
 import { expect } from '@storybook/jest';
 import { userEvent } from '@storybook/testing-library';
 import SecurityNumberInput from '.';
-import useSecurityNumber from '../../hooks/useSecurityNumber';
-import MESSAGE from '../../constant/message';
+import useSecurityNumber from '../../../hooks/useSecurityNumber';
+import MESSAGE from '../../../constant/message';
 import { CLASS } from 'constant/selector';
 
 export default {
-  title: 'Payment/SecurityNumberInput',
+  title: 'Molecules/SecurityNumberInput',
   component: SecurityNumberInput,
   argTypes: {
     securityNumber: {
