+++ src/pages/CardAddPage/index.jsx
@@ -1,14 +1,14 @@
 import { useNavigate } from 'react-router-dom';
-import Head from '../../components/Head';
-import Card from '../../components/Card';
-import CardNumbersInput from '../../components/CardNumbersInput';
-import ExpiredDateInput from '../../components/ExpiredDateInput';
-import OwnerNameInput from '../../components/OwnerNameInput';
-import SecurityNumberInput from '../../components/SecurityNumberInput';
-import PasswordInput from '../../components/PasswordInput';
-import Button from '../../components/Button';
-import ModalPortal from '../../components/ModalPortal';
-import CardCompanySelector from '../../components/CardCompanySelector';
+import Button from '../../components/Atoms/Button';
+import Card from '../../components/Molecules/Card';
+import CardNumbersInput from '../../components/Molecules/CardNumbersInput';
+import ExpiredDateInput from '../../components/Molecules/ExpiredDateInput';
+import OwnerNameInput from '../../components/Molecules/OwnerNameInput';
+import SecurityNumberInput from '../../components/Molecules/SecurityNumberInput';
+import PasswordInput from '../../components/Molecules/PasswordInput';
+import CardCompanySelector from '../../components/Molecules/CardCompanySelector';
+import Head from '../../components/Templates/Head';
+import ModalPortal from '../../components/Templates/ModalPortal';
 import { Page, HeadContainer, BackButton, CardSection, Form, SubmitButtonContainer } from './style';
 import MESSAGE from '../../constant/message';
 import { useCardListContext } from 'context/cardList';
